requirements.txt
streamlit
numpy
pandas
pillow
scipy